deleted some files
Almost finished. Just need to set up the tableViews
