+++ NextDorm.xcworkspace/xcuserdata/daisyrocha.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "5824C932-89A2-459A-BE63-1D1B4CF68795"
+   type = "0"
+   version = "2.0">
+</Bucket>
